+++ src/bin/oauth/login/post.rs
@@ -1,8 +1,8 @@
 use cookie::Cookie;
 use http::{HeaderMap, HeaderValue};
 use lambda_http::{run, service_fn, Error, IntoResponse, Request, RequestExt};
-use oauth_flow::models::user::User;
-use oauth_flow::queries::get_user_query::{GetUser, GetUserRequest};
+use oauth_flow::dtos::login::get_user_request::GetUserRequest;
+use oauth_flow::queries::get_user_query::GetUser;
 use oauth_flow::setup_tracing;
 use oauth_flow::utils::api_helper::{ApiResponseType, IsCors};
 use oauth_flow::utils::cookie::CookieExt;
@@ -38,38 +38,33 @@ pub async fn handler(
     event: Request,
 ) -> Result<impl IntoResponse, Error> {
     println!("{event:?}");
-    let user = login(app_client, &event)
-        .await
-        .ok()
-        .and_then(|result| result);
+    let request = GetUserRequest::validate(&event);
+    if let Some(request) = request {
+        let user = app_client.query(&request).await.ok().and_then(|user| user);
+        if let Some(user) = user {
+            let cookie = Cookie::to_cookie_string(
+                String::from("myOAuth"),
+                HashMap::from([
+                    (String::from("email"), user.email.unwrap()),
+                    (
+                        String::from("is_consent"),
+                        user.is_consent.unwrap().to_string(),
+                    ),
+                    (String::from("is_optin"), user.is_optin.unwrap().to_string()),
+                ]),
+            );
+            let mut headers = HeaderMap::new();
+            headers.insert(http::header::SET_COOKIE, HeaderValue::from_str(&cookie)?);
 
-    let host = event
-        .headers()
-        .get("Host");
+            let target = ApiResponseType::build_url_from_querystring(
+                format!("https://{}{}", request.host, app_client.redirect_path()),
+                event.query_string_parameters(),
+            );
 
-    if let (Some(user), Some(host)) = (user, host) {
-        let cookie = Cookie::to_cookie_string(
-            String::from("myOAuth"),
-            HashMap::from([
-                (String::from("email"), user.email.unwrap()),
-                (
-                    String::from("is_consent"),
-                    user.is_consent.unwrap().to_string(),
-                ),
-                (String::from("is_optin"), user.is_optin.unwrap().to_string()),
-            ]),
-        );
-        let mut headers = HeaderMap::new();
-        headers.insert(http::header::SET_COOKIE, HeaderValue::from_str(&cookie)?);
-
-        let target = ApiResponseType::build_url_from_querystring(
-            format!("https://{}{}", host.to_str()?, app_client.redirect_path()),
-            event.query_string_parameters(),
-        );
-
-        return Ok(
-            ApiResponseType::FoundWithCustomHeaders(target, IsCors::No, headers).to_response(),
-        );
+            return Ok(
+                ApiResponseType::FoundWithCustomHeaders(target, IsCors::No, headers).to_response(),
+            );
+        }
     }
 
     Ok(ApiResponseType::NoContent(
@@ -77,24 +72,4 @@ pub async fn handler(
         IsCors::No,
     )
     .to_response())
-}
-
-async fn login(
-    app_client: &dyn PostAppInitialisation,
-    event: &Request,
-) -> Result<Option<User>, Error> {
-    let user = event.payload::<User>()?.unwrap();
-    let query_params = event.query_string_parameters();
-    let client_id = query_params
-        .first("client_id")
-        .expect("client_id not found");
-
-    let request = GetUserRequest::builder()
-        .client_id(client_id)
-        .email(user.email.unwrap())
-        .build();
-
-    let user = app_client.query(&request).await?;
-
-    Ok(user)
 }
