+++ src/dtos/mod.rs
@@ -1,2 +1,3 @@
 pub mod consent;
 pub mod optin;
+pub mod login;
