+++ src/dtos/optin/mod.rs
@@ -1,1 +1,1 @@
-pub mod page_request;
+pub mod update_optIn_request;
