+++ src/utils/injections/oauth/consent/post_di.rs
@@ -1,6 +1,6 @@
 use crate::{
     error::ApplicationError,
-    queries::update_consent_query::{UpdateConsent, UpdateConsentQuery}, dtos::consent::page_request::UpdateConsentRequest,
+    queries::update_consent_query::{UpdateConsent, UpdateConsentQuery}, dtos::consent::update_consent_request::UpdateConsentRequest,
 };
 use async_trait::async_trait;
 use typed_builder::TypedBuilder as Builder;
